setup cmake (#469)
* set up cmake for dev in windows

* add git link

* remove zlib as it's not needed.

* fix most of the comments.

* change cmake version to 3.1

* remove readme

* add pybind11 into cmake

* use local protobuf.

* remove include for protobuf as it's not needed.

* merge all things in Cmakelists.txt in root folder

* update Cmakelists.txt in root folder

* use customized generation function for proto header/code generation.

* adding comments.

+++ onnx/common/ir_pb_converter.cc
@@ -374,13 +374,13 @@ void addAttribute(ONNX_NAMESPACE::NodeProto * n_p, Node * n, Symbol name) {
   attr->set_name(name.toString());
   switch(n->kindOf(name)) {
     case AttributeKind::f:
-      attr->set_f(n->f(name));
+      attr->set_f(static_cast<float>(n->f(name)));
       attr->set_type(ONNX_NAMESPACE::AttributeProto_AttributeType_FLOAT);
       break;
     case AttributeKind::fs:
       attr->set_type(ONNX_NAMESPACE::AttributeProto_AttributeType_FLOATS);
       for(auto & v : n->fs(name))
-        attr->add_floats(v);
+        attr->add_floats(static_cast<float>(v));
       break;
     case AttributeKind::i:
       attr->set_type(ONNX_NAMESPACE::AttributeProto_AttributeType_INT);
